startMenu reorganize
reordered some option on the menu explorer in unity itself. I also readjusted the positioning of some of the text. the main screen buttons' text should be flush and should be uniformly distanced now. removed my middle name from the credits so that my name doesnt overlap with the logo.